+++ src/screens/book/BookRead/index.js
@@ -14,18 +14,19 @@ import {
 import {SafeAreaView} from 'react-native-safe-area-context';
 import {useFocusEffect, useRoute} from '@react-navigation/native';
 import Ionic from 'react-native-vector-icons/Ionicons';
-import ProgressBar from '../components/book/ProgressBar.js';
-import SettingModal from '../components/modals/SettingModal.js';
-import YesNoModal from '../components/common/YesNoModal.js';
-import NextStep from '../components/book/NextStep.js';
-import BookBg from '../../assets/images/bookBg.png';
-import fetchWithAuth from '../api/fetchWithAuth.js';
+import ProgressBar from '../../../components/book/ProgressBar.js';
+import SettingModal from '../../../components/modals/SettingModal.js';
+import YesNoModal from '../../../components/common/YesNoModal.js';
+import NextStep from '../../../components/book/NextStep.js';
+import BookBg from '../../../../assets/images/bookBg.png';
+import fetchWithAuth from '../../../api/fetchWithAuth.js';
 import Tts from 'react-native-tts';
 import {WebView} from 'react-native-webview';
+import { styles } from './styles';
 
 const BookRead = ({navigation}) => {
   const route = useRoute();
-  const {profileId, bookId} = route.params; // 파라미터 받기
+  const {profileId, bookId} = route.params;
 
   const [title, setTitle] = useState('');
   const [currentPage, setCurrentPage] = useState(1);
@@ -133,7 +134,7 @@ const BookRead = ({navigation}) => {
         const initialPage = result.data.currentPage + 1;
         setCurrentPage(initialPage);
         await fetchPageDetails(initialPage);
-        setIsLoading(false); // 로딩 완료
+        setIsLoading(false);
       } else {
         Alert.alert('Error', '책 세부정보 불러오기 실패');
       }
@@ -221,7 +222,7 @@ const BookRead = ({navigation}) => {
     };
 
     Tts.setDefaultLanguage('en-US');
-    setTtsRate(initialSpeed); // initialSpeed를 기반으로 속도 설정
+    setTtsRate(initialSpeed);
     Tts.setDefaultPitch(1.0);
 
     Tts.voices().then(voices => {
@@ -589,125 +590,4 @@ const BookRead = ({navigation}) => {
   );
 };
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#FBF7EC',
-  },
-  page: {
-    width: '70.5%',
-    height: '100%',
-  },
-
-  titleBox: {
-    marginBottom: 20,
-  },
-  textBox: {
-    width: '35%',
-    justifyContent: 'center',
-    position: 'absolute',
-    top: '3%',
-    left: '57%',
-    borderRadius: 10,
-    padding: 10,
-  },
-  bookTextContainer: {
-    flexDirection: 'row',
-    flexWrap: 'wrap',
-    justifyContent: 'center',
-  },
-  bookTitle: {
-    fontSize: 38,
-    fontWeight: 'bold',
-    color: 'black',
-    textAlign: 'center',
-    fontFamily: 'TAEBAEKfont',
-  },
-  iconBox: {
-    flexDirection: 'column',
-    justifyContent: 'space-around',
-    alignItems: 'baseline',
-    margin: 20,
-  },
-  icon: {
-    marginTop: 10,
-    backgroundColor: 'white',
-    borderRadius: 10,
-  },
-  bookText: {
-    fontSize: 18,
-    color: '#000000',
-    textAlign: 'center',
-    lineHeight: 30,
-    fontFamily: 'TAEBAEKfont',
-  },
-  wordBox: {
-    width: 150,
-    height: 70,
-    position: 'absolute',
-    bottom: '10%',
-    left: '80%',
-    borderTopWidth: 1,
-    borderBottomWidth: 1,
-    borderColor: 'rgba(78,90,140,0.8)',
-    borderStyle: 'solid',
-    justifyContent: 'center',
-  },
-  wordText: {
-    textAlign: 'center',
-    fontSize: 18,
-    color: 'rgba(78,90,140,0.8)',
-    fontFamily: 'TAEBAEKmilkyway',
-  },
-  highlightedText: {
-    backgroundColor: 'yellow',
-  },
-  modalBackground: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  highlightModal: {
-    position: 'absolute',
-    padding: 10,
-    backgroundColor: 'white',
-    borderRadius: 10,
-    alignItems: 'center',
-  },
-  highlightModalButtons: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-  },
-  highlightModalText: {
-    paddingHorizontal: 10,
-    fontFamily: 'TAEBAEKmilkyway',
-  },
-  emptyContainer: {
-    flex: 1,
-  },
-  loadingContainer: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  modalContainer: {
-    flex: 1,
-    backgroundColor: 'rgba(0, 0, 0, 0.5)',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  webviewContainer: {
-    width: '80%',
-    height: '90%',
-    backgroundColor: 'white',
-    borderRadius: 10,
-    overflow: 'hidden',
-  },
-  closeButton: {
-    padding: 5,
-    paddingHorizontal: 15,
-    backgroundColor: '#fff',
-    alignItems: 'flex-end',
-  },
-});
 export default BookRead;
