+++ src/screens/book/BookShelf/styles.js
@@ -1,0 +1,137 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'flex-start',
+    paddingTop: 30,
+    backgroundColor: '#FBF7EC',
+  },
+  radioContainer: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    width: '50%',
+    marginBottom: 25,
+    paddingTop: 10,
+  },
+  radioButton: {
+    flex: 1,
+    marginHorizontal: 10,
+    paddingVertical: 12,
+    alignItems: 'center',
+  },
+  radioButtonAll: {
+    marginRight: -35,
+    paddingHorizontal: 15,
+    width: 90,
+    height: 45,
+  },
+  radioButtonFavorite: {
+    marginLeft: -25,
+    paddingHorizontal: 15,
+    width: 180,
+    height: 45,
+  },
+  radioButtonReading: {
+    marginLeft: -65,
+    paddingHorizontal: 15,
+    width: 170,
+    height: 45,
+  },
+  radioButtonGradient: {
+    borderRadius: 25,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  selected: {
+    borderWidth: 0,
+    borderColor: '#FF8C43',
+    shadowColor: '#000',
+    shadowOffset: {width: 0, height: 2},
+    shadowOpacity: 0.3,
+    shadowRadius: 3,
+  },
+  radioButtonText: {
+    color: '#C3C3C3',
+    fontWeight: 'bold',
+    fontSize: 30,
+  },
+  radioButtonTextAll: {},
+  radioButtonTextFavorite: {},
+  radioButtonTextReading: {},
+  radioButtonTextSelected: {
+    color: '#000',
+  },
+  shelfWrapper: {
+    flex: 1,
+    justifyContent: 'flex-start',
+    width: '100%',
+    marginTop: 100,
+  },
+  shelfContainer: {
+    width: '100%',
+    marginBottom: 70,
+    position: 'relative',
+  },
+  shelf: {
+    width: '100%',
+    height: 100,
+    resizeMode: 'contain',
+    top: 20,
+  },
+  bookButton: {
+    position: 'absolute',
+    top: -68,
+  },
+  bookImage: {
+    width: 140,
+    height: 130,
+    resizeMode: 'contain',
+  },
+  favoriteButton: {
+    position: 'absolute',
+    bottom: 5,
+    left: 5,
+    backgroundColor: 'transparent',
+  },
+  squareButton: {
+    position: 'absolute',
+    top: 20,
+    right: 20,
+    width: 110,
+    height: 110,
+    borderRadius: 30,
+    backgroundColor: '#EFEFEF',
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  squareButtonImage: {
+    width: '100%',
+    height: '100%',
+    resizeMode: 'contain',
+    borderRadius: 30,
+  },
+  roundButton: {
+    position: 'absolute',
+    bottom: 5,
+    alignSelf: 'center',
+    width: 80,
+    height: 80,
+    borderRadius: 40,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  roundButtonGradient: {
+    width: '100%',
+    height: '100%',
+    borderRadius: 40,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  roundButtonImage: {
+    width: 40,
+    height: 40,
+    resizeMode: 'contain',
+  },
+});
