+++ src/screens/profile/Profile/styles.js
@@ -1,0 +1,108 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#FBF7EC',
+    alignItems: 'center',
+    justifyContent: 'flex-start',
+  },
+  header: {
+    fontSize: 37,
+    marginBottom: 50,
+    marginTop: 160,
+    color: '#393939',
+    fontWeight: '800',
+  },
+  profilesContainer: {
+    flexDirection: 'row',
+    flexWrap: 'wrap',
+    alignItems: 'flex-start',
+    justifyContent: 'center',
+    marginTop: 30,
+  },
+  profileContainer: {
+    alignItems: 'center',
+    margin: 10,
+  },
+  profileButton: {
+    width: 150,
+    height: 150,
+    backgroundColor: 'lightblue',
+    borderRadius: 20,
+    alignItems: 'center',
+    justifyContent: 'center',
+    position: 'relative',
+    overflow: 'hidden',
+    shadowColor: '#000',
+    shadowOffset: {width: 0, height: 2},
+    shadowOpacity: 0.3,
+    shadowRadius: 5,
+    elevation: 5,
+  },
+  profileButtonActive: {
+    opacity: 0.6,
+  },
+  profileImage: {
+    width: '110%',
+    height: '114%',
+    resizeMode: 'cover',
+  },
+  profileText: {
+    fontSize: 20,
+    fontWeight: '700',
+    color: '#545454',
+    textAlign: 'center',
+    marginTop: 5,
+  },
+  addButton: {
+    width: 150,
+    height: 150,
+    backgroundColor: '#FCAE59',
+    borderRadius: 20,
+    alignItems: 'center',
+    justifyContent: 'center',
+    margin: 10,
+    shadowColor: '#000',
+    shadowOffset: {width: 0, height: 2},
+    shadowOpacity: 0.3,
+    shadowRadius: 5,
+    elevation: 5,
+  },
+  plus: {
+    fontSize: 110,
+    fontWeight: 'bold',
+    color: '#FF8B42',
+    textAlign: 'center',
+  },
+  changeProfileButton: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    position: 'absolute',
+    bottom: 20,
+    backgroundColor: 'transparent',
+    paddingHorizontal: 20,
+    paddingVertical: 10,
+    borderRadius: 10,
+  },
+  changeProfileButtonText: {
+    fontSize: 18,
+    fontWeight: '900',
+  },
+  penIcon: {
+    width: 24,
+    height: 24,
+    tintColor: '#393939',
+  },
+  overlay: {
+    ...StyleSheet.absoluteFillObject,
+    backgroundColor: 'rgba(0, 0, 0, 0.4)',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  profilepenIcon: {
+    width: 70,
+    height: 70,
+    tintColor: '#FFFFFF',
+  },
+});
